rename ArchitectureTest to *Sniffer
