+++ Kununu/ArchitectureSniffer/Configuration/Rules/MustOnlyDependOnWhitelist.php
@@ -1,11 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace Kununu\ArchitectureTest\Configuration\Rules;
+namespace Kununu\ArchitectureSniffer\Configuration\Rules;
 
 use JsonException;
-use Kununu\ArchitectureTest\Configuration\Selector\Selectable;
-use Kununu\ArchitectureTest\Configuration\Selectors;
+use Kununu\ArchitectureSniffer\Configuration\Selector\Selectable;
+use Kununu\ArchitectureSniffer\Configuration\Selectors;
 use PHPat\Selector\Selector;
 use PHPat\Test\PHPat;
