+++ Kununu/ArchitectureSniffer/Configuration/Selector/ClassSelector.php
@@ -1,16 +1,16 @@
 <?php
 declare(strict_types=1);
 
-namespace Kununu\ArchitectureTest\Configuration\Selector;
+namespace Kununu\ArchitectureSniffer\Configuration\Selector;
 
 use PHPat\Selector\Selector;
 use PHPat\Selector\SelectorInterface;
 
-final readonly class InterfaceClassSelector implements Selectable
+final readonly class ClassSelector implements Selectable
 {
     use RegexTrait;
 
-    public const string KEY = 'interface';
+    public const string KEY = 'class';
 
     public function __construct(
         public string $name,
@@ -22,10 +22,7 @@ final readonly class InterfaceClassSelector implements Selectable
     {
         $namespace = $this->makeRegex($this->namespace);
 
-        return Selector::AllOf(
-            Selector::classname($namespace, $namespace !== $this->namespace),
-            Selector::isInterface(),
-        );
+        return Selector::classname($namespace, $namespace !== $this->namespace);
     }
 
     public function getName(): string
